refactor!: separate raillabel format into its own directory
