+++ raillabel/format/__init__.py
@@ -2,26 +2,26 @@
 # SPDX-License-Identifier: Apache-2.0
 """Module containing all relevant OpenLABEL format classes."""
 
-from ._object_annotation import _ObjectAnnotation, annotation_classes
-from .bbox import Bbox
-from .cuboid import Cuboid
-from .element_data_pointer import ElementDataPointer
-from .frame import Frame
-from .frame_interval import FrameInterval
-from .intrinsics_pinhole import IntrinsicsPinhole
-from .intrinsics_radar import IntrinsicsRadar
-from .metadata import Metadata
-from .num import Num
-from .object import Object
-from .point2d import Point2d
-from .point3d import Point3d
-from .poly2d import Poly2d
-from .poly3d import Poly3d
-from .quaternion import Quaternion
-from .scene import Scene
-from .seg3d import Seg3d
-from .sensor import Sensor, SensorType
-from .sensor_reference import SensorReference
-from .size2d import Size2d
-from .size3d import Size3d
-from .transform import Transform
+from .raillabel._object_annotation import _ObjectAnnotation, annotation_classes
+from .raillabel.bbox import Bbox
+from .raillabel.cuboid import Cuboid
+from .raillabel.element_data_pointer import ElementDataPointer
+from .raillabel.frame import Frame
+from .raillabel.frame_interval import FrameInterval
+from .raillabel.intrinsics_pinhole import IntrinsicsPinhole
+from .raillabel.intrinsics_radar import IntrinsicsRadar
+from .raillabel.metadata import Metadata
+from .raillabel.num import Num
+from .raillabel.object import Object
+from .raillabel.point2d import Point2d
+from .raillabel.point3d import Point3d
+from .raillabel.poly2d import Poly2d
+from .raillabel.poly3d import Poly3d
+from .raillabel.quaternion import Quaternion
+from .raillabel.scene import Scene
+from .raillabel.seg3d import Seg3d
+from .raillabel.sensor import Sensor, SensorType
+from .raillabel.sensor_reference import SensorReference
+from .raillabel.size2d import Size2d
+from .raillabel.size3d import Size3d
+from .raillabel.transform import Transform
