+++ raillabel/format/raillabel/frame.py
@@ -6,7 +6,7 @@ import typing as t
 import uuid
 from dataclasses import dataclass, field
 
-from .._util._warning import _warning
+from ..._util._warning import _warning
 from ._object_annotation import _ObjectAnnotation, annotation_classes
 from .num import Num
 from .object import Object
