+++ raillabel/format/raillabel/num.py
@@ -4,7 +4,7 @@
 import typing as t
 from dataclasses import dataclass
 
-from .._util._warning import _warning
+from ..._util._warning import _warning
 from .sensor import Sensor
