+++ tests/test_raillabel/format/raillabel/test_bbox.py
@@ -10,7 +10,7 @@ import pytest
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
 from raillabel._util._warning import _WarningsLogger
-from raillabel.format.bbox import Bbox
+from raillabel.format import Bbox
 
 # == Fixtures =========================
