+++ tests/test_raillabel/format/raillabel/test_frame.py
@@ -11,7 +11,7 @@ import pytest
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
 from raillabel._util._warning import _WarningsLogger
-from raillabel.format.frame import Frame
+from raillabel.format import Frame
 
 # == Fixtures =========================
