+++ tests/test_raillabel/format/raillabel/test_num.py
@@ -9,7 +9,7 @@ import pytest
 
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
-from raillabel.format.num import Num
+from raillabel.format import Num
 
 # == Fixtures =========================
