+++ tests/test_raillabel/format/raillabel/test_poly2d.py
@@ -9,7 +9,7 @@ import pytest
 
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
-from raillabel.format.poly2d import Poly2d
+from raillabel.format import Poly2d
 
 # == Fixtures =========================
