+++ tests/test_raillabel/format/raillabel/test_seg3d.py
@@ -9,7 +9,7 @@ import pytest
 
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
-from raillabel.format.seg3d import Seg3d
+from raillabel.format import Seg3d
 
 # == Fixtures =========================
